+++ examples/eval.rs
@@ -1,18 +1,18 @@
 #![feature(box_patterns)]
 
-use collect_mac::collect;
-use indexmap::IndexMap as Map;
+
+
 use partiql::parser;
 use partiql::pqlir_parser;
 use partiql::sql::evaluate;
-use partiql::sql::re_from_str;
-use partiql::sql::restrict;
-use partiql::sql::Bindings;
-use partiql::sql::DPath;
-use partiql::sql::Expr;
-use partiql::sql::Field;
-use partiql::sql::WhereCond;
-use partiql::value::PqlValue;
+
+
+
+
+
+
+
+
 
 fn main() -> anyhow::Result<()> {
     let input = "
