+++ examples/pq.rs
@@ -1,5 +1,5 @@
-use std::io::{self, Read};
-use std::path::PathBuf;
+
+
 use std::str::FromStr;
 
 use partiql::lang::{Lang, LangType};
