+++ src/lib/sql/field.rs
@@ -1,7 +1,7 @@
-use collect_mac::collect;
+
 
 use crate::sql::Bindings;
-use crate::sql::Expr;
+
 use std::collections::VecDeque;
 
 #[derive(Debug, Default, Clone, PartialEq)]
