+++ src/lib/sql/mod.rs
@@ -1,4 +1,4 @@
-use crate::value::PqlValue;
+
 
 mod bindings;
 mod eval;
