+++ src/lib/value/pql_vector.rs
@@ -1,10 +1,10 @@
-use std::collections::{BTreeMap, BTreeSet};
+
 use std::ops::{Add, Div, Mul, Neg, Rem, Sub};
 
-use indexmap::IndexMap;
-use ordered_float::OrderedFloat;
+
+
 use rayon::prelude::*;
-use serde_derive::{Deserialize, Serialize};
+
 
 use crate::value::PqlValue;
