+++ tests/math.rs
@@ -1,5 +1,5 @@
 use std::convert::TryFrom;
-use std::convert::TryInto;
+
 
 use partiql::parser;
 use partiql::value::PqlValue;
